feat(agent): the spend axis spans a task, not one Run
A per-Run total cannot see the failure this axis exists for. runLoopState is
rebuilt every Run, so every "continue" restarts the count from zero — and
the reported runaway was four hours, which is many Runs. A ceiling or a
budget scoped to one Run offers exactly no protection against it.

The task scope resets with the evidence ledger. That boundary already means
"this is new work": a continuation preserves the ledger, an unrelated turn
resets it, and the delivery scope decides which. Reusing it keeps one answer
to "is this the same task" instead of inventing a second one that can
disagree with the first.

Samples now carry both scopes. Turn is what this Run spent, Task is what the
work has cost since it started; a reader that wants either has it without
differencing.

Still shadow — no threshold reads it.

agent.go was 4 lines over its file-size budget with the new field, so the
storm breaker moved to storm_breaker.go: threshold, block message, signature,
and the guard itself were already one concern sitting in the middle of
another file. agent.go drops 151 lines and 13 essay findings, which the
baseline diff moves rather than forgives.
