+++ internal/event/run_budget.go
@@ -2,18 +2,26 @@ package event
 
 import "reasonix/internal/nilutil"
 
-// RunBudgetSample is one turn's accumulated spend after a round: counts and
-// money, never content. Priced is false when no price table covered the turn,
-// so a reader can tell a genuinely free turn from an unpriced one.
-type RunBudgetSample struct {
+// RunBudgetTotals is one scope's accumulated spend: counts and money, never
+// content. Priced is false when no price table covered it, so a reader can
+// tell a genuinely free stretch from an unpriced one.
+type RunBudgetTotals struct {
 	Rounds       int     `json:"rounds"`
 	Requests     int     `json:"requests"`
 	PromptTokens int     `json:"promptTokens"`
 	OutputTokens int     `json:"outputTokens"`
 	Cost         float64 `json:"cost"`
-	Currency     string  `json:"currency,omitempty"`
 	Priced       bool    `json:"priced"`
 	ElapsedMs    int64   `json:"elapsedMs"`
+}
+
+// RunBudgetSample reports both scopes after a round. Turn is the current Run;
+// Task spans every Run continuing the same work, because "continue" starts a
+// new Run and the spend worth stopping accrues across them.
+type RunBudgetSample struct {
+	Turn     RunBudgetTotals `json:"turn"`
+	Task     RunBudgetTotals `json:"task"`
+	Currency string          `json:"currency,omitempty"`
 }
 
 // RunBudgetSink is an optional sink capability for the per-round spend axis.
